+++ core-ng/src/main/java/core/framework/impl/log/LogEvent.java
@@ -2,6 +2,7 @@ package core.framework.impl.log;
 
 import core.framework.impl.log.marker.ErrorCodeMarker;
 import core.framework.util.Exceptions;
+import core.framework.util.Network;
 import org.slf4j.Marker;
 
 import java.time.Instant;
@@ -46,7 +47,7 @@ final class LogEvent {
         var now = Instant.now();
         var builder = new StringBuilder(256);
         builder.append(DateTimeFormatter.ISO_INSTANT.format(now))
-               .append(" [").append(Thread.currentThread().getName()).append("] ")
+               .append(" [").append(Network.LOCAL_HOST_NAME).append("] [").append(Thread.currentThread().getName()).append("] ")
                .append(level.name())
                .append(' ')
                .append(logger)