+++ core-ng/src/main/java/core/framework/http/HTTPClientBuilder.java
@@ -35,7 +35,8 @@ public final class HTTPClientBuilder {
 
     private final Logger logger = LoggerFactory.getLogger(HTTPClientBuilder.class);
 
-    private Duration timeout = Duration.ofSeconds(60);
+    private Duration connectTimeout = Duration.ofSeconds(15);
+    private Duration requestTimeout = Duration.ofSeconds(60);
     private Duration slowOperationThreshold = Duration.ofSeconds(30);
     private boolean enableCookie = false;
     private boolean enableRedirect = false;
@@ -54,12 +55,12 @@ public final class HTTPClientBuilder {
                 builder.sslContext(sslContext);
             }
 
-            builder.connectTimeout(Duration.ofMillis(timeout.toMillis() / 2));  // use shorter connect timeout for fast fail
+            builder.connectTimeout(connectTimeout);
             if (enableRedirect) builder.followRedirects(HttpClient.Redirect.NORMAL);
             if (enableCookie) builder.cookieHandler(new CookieManager(null, CookiePolicy.ACCEPT_ALL));
             HttpClient httpClient = builder.build();
 
-            return new HTTPClientImpl(httpClient, userAgent, timeout, maxRetries, slowOperationThreshold);
+            return new HTTPClientImpl(httpClient, userAgent, requestTimeout, maxRetries, slowOperationThreshold);
         } catch (NoSuchAlgorithmException | KeyManagementException e) {
             throw new Error(e);
         } finally {
@@ -67,8 +68,9 @@ public final class HTTPClientBuilder {
         }
     }
 
-    public HTTPClientBuilder timeout(Duration timeout) {
-        this.timeout = timeout;
+    public HTTPClientBuilder timeout(Duration connectTimeout, Duration requestTimeout) {
+        this.connectTimeout = connectTimeout;
+        this.requestTimeout = requestTimeout;
         return this;
     }