+++ core-ng/src/main/java/core/framework/internal/log/marker/ErrorCodeMarker.java
@@ -1,4 +1,4 @@
-package core.framework.impl.log.marker;
+package core.framework.internal.log.marker;
 
 /**
  * @author neo